# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TuTramiteFacil API",
        "description": "API REST para gestión de solicitantes y solicitudes de ayudas sociales (DDD + Clean Architecture).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8010",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/solicitantes": {
            "get": {
                "tags": [
                    "Solicitantes"
                ],
                "summary": "Listar todos los solicitantes",
                "description": "GET /api/solicitantes\nPúblico (sin autenticación), según el enunciado.",
                "operationId": "313d91180355d1417ae35759a4ed1711",
                "responses": {
                    "200": {
                        "description": "Lista de solicitantes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Solicitantes"
                ],
                "summary": "Crear un nuevo solicitante",
                "description": "POST /api/solicitantes\nRequiere JWT (se configurará en las rutas).",
                "operationId": "765cc3a5cf6beda0ab59a4a62a8214e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellidos",
                                    "email",
                                    "telefono",
                                    "comunidad_autonoma",
                                    "fecha_registro"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "Macero Delgado"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "600000000"
                                    },
                                    "comunidad_autonoma": {
                                        "type": "string",
                                        "example": "Comunidad Valenciana"
                                    },
                                    "fecha_registro": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitante creado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/solicitantes/{solicitante}": {
            "get": {
                "tags": [
                    "Solicitantes"
                ],
                "summary": "Ver detalle de un solicitante",
                "description": "GET /api/solicitantes/{solicitante}\nPúblico.",
                "operationId": "40dd1d9b0efaf7c641679f6b92408958",
                "parameters": [
                    {
                        "name": "solicitante",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del solicitante"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Solicitantes"
                ],
                "summary": "Actualizar un solicitante",
                "description": "PUT /api/solicitantes/{solicitante}\nRequiere JWT.",
                "operationId": "d7a77d14947c526d085cf5314e87fe18",
                "parameters": [
                    {
                        "name": "solicitante",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "Macero Delgado"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "600000000"
                                    },
                                    "comunidad_autonoma": {
                                        "type": "string",
                                        "example": "Madrid"
                                    },
                                    "fecha_registro": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Solicitante actualizado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Solicitantes"
                ],
                "summary": "Eliminar un solicitante",
                "description": "DELETE /api/solicitantes/{solicitante}\nRequiere JWT.",
                "operationId": "6f19539874ca2eef321b36ecaf7f15c2",
                "parameters": [
                    {
                        "name": "solicitante",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}